2D Physics (#53)
* WIP

* Add box2d submodule

* WIP

* Modify box2d

* Remove original box2d

* Add forked version of box2d as submodule

* WIP

+++ editor/source/editor_camera.hpp
@@ -21,7 +21,7 @@ namespace jng {
         void onUpdate();
         void onEvent(Event& event);
 
-        void setViewportSize(uint32 width, uint32 height);
+        void setViewportSize(float width, float height);
 
         glm::mat4 getVP() const { return Camera::getVP(m_transform); }
     private:
